+++ packages/data-explorer/src/components/MetadataWarning.tsx
@@ -1,0 +1,16 @@
+const MetadataWarning = ({ metadata }: { metadata: Metadata }) => {
+  const warning =
+    metadata && metadata.sampled ? (
+      <span>
+        <b>NOTE:</b> This data is sampled
+      </span>
+    ) : null;
+
+  return (
+    <MetadataWarningWrapper>
+      {warning ? (
+        <MetadataWarningContent>{warning}</MetadataWarningContent>
+      ) : null}
+    </MetadataWarningWrapper>
+  );
+};
